+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/Checkout/CheckoutApplePayContextFunctionalTest.swift
@@ -94,7 +94,8 @@ class CheckoutApplePayContextFunctionalTest: STPNetworkStubbingTestCase {
         let context = CheckoutApplePayContext(
             checkoutSession: session,
             applePayConfirmationParameters: applePayConfirmationParameters,
-            authorizationController: MockPKPaymentAuthorizationController()
+            authorizationController: MockPKPaymentAuthorizationController(),
+            checkoutWalletUpdater: MockCheckoutSessionWalletUpdater()
         )
         return (context, apiClient)
     }