+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodEnums.swift
@@ -104,6 +104,10 @@ import Foundation
     case bizum
     /// A Korean cards payment method
     case krCard
+    /// A Naver Pay payment method
+    case naverPay
+    /// A PAYCO payment method
+    case payco
     /// An unknown type.
     case unknown
 
@@ -206,6 +210,10 @@ import Foundation
             return "Bizum"
         case .krCard:
             return "Korean cards"
+        case .naverPay:
+            return "Naver Pay"
+        case .payco:
+            return "PAYCO"
         case .cardPresent,
             .unknown:
             return STPLocalizedString("Unknown", "Default missing source type label")
@@ -311,6 +319,10 @@ import Foundation
             return "bizum"
         case .krCard:
             return "kr_card"
+        case .naverPay:
+            return "naver_pay"
+        case .payco:
+            return "payco"
         }
     }